// src/app/models/priv-msg.model.ts
export interface PrivMsgModel {
  badgeInfo?: string;
  badges?: string;
  clientNonce?: string;
  color?: string;
  displayName?: string;
  emotes?: string;
  flags?: string;
  id?: string;
  mod?: number;
  roomId?: string;
  subscriber?: string;
  tmiSentTs?: string;
  turbo?: string;
  userId?: string;
  userType?: string;
  raw: string;
}